% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capabilities.R
\name{capabilities}
\alias{capabilities}
\title{Choose how extensively suggests affects a package namespace}
\usage{
capabilities(name = c("colon", "create"), env = parent.frame())
}
\description{
\code{suggests} will invariably inject some objects into the package namespace.
However, the extent to which this happens may be undesirable. At a minimum,
\code{suggests} will will add a \code{.suggests} object to your package namespace. From
there, \code{suggests} can mask \code{base} operators in your namespace to streamline
your ' development experience.
}
\details{
To enable these extra capabilities, you'll have to opt-in to them using
\code{\link[=capabilities]{capabilities()}}.
}
\section{Capabilitiess}{

\describe{
\item{"colon"}{
Mask base \code{::} and \code{:::} internally within your package namespace. The
masking functions will apply your specified behavior when the namespace
is attempted to be accessed.
}
\item{"create"}{
Creates an object of the packages name in the calling namespace '
automatically. It can be indexed just like a package namespace, using
the \code{::} and \code{:::} operators (if feature "colon" is enabled), or
using \code{$}.
}
}
}

\examples{
\dontrun{capabilities("colon")
utils <- suggests("utils")
utils::head(1:10)

capabilities("create")
suggests("not.a.real.package")
not.a.real.package::head(1:10)
# Error: This feature is unavailable because package 'not.a.real.package'
#   is not installed.
}

}
