% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_installed.R
\name{check_installed}
\alias{check_installed}
\title{Check whether packages are installed}
\usage{
check_installed(pkgs, load = TRUE, lib.loc = .libPaths())
}
\arguments{
\item{pkgs}{Character vector of package names to check.}

\item{load}{Whether to make sure the packages can be loaded.}

\item{lib.loc}{Library locations to check, \code{\link[=.libPaths]{.libPaths()}} by default.}
}
\description{
If \code{load = TRUE}, try to load the namespace of each package in turn; this is
slower, but less likely to return incorrect results.
}
\details{
If \code{load = FALSE}, check whether a folder for the package exists in any of
the specified libraries. Much faster, but less robust!
}
