% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colon_ops.R
\name{colon-ops}
\alias{colon-ops}
\alias{suggests_double_colon}
\alias{suggests_triple_colon}
\title{Replacement \code{::} and \code{:::} operators}
\usage{
suggests_double_colon(pkg, name)

suggests_triple_colon(pkg, name)
}
\arguments{
\item{pkg}{package name: symbol or literal character string.}

\item{name}{variable name: symbol or literal character string.}
}
\description{
The only difference in behavior from the \code{base} equivalents is that the
left-hand-side is first checked to see if it is a suggested package, which
prompts \code{suggests} to pull registered data about the package from the
encapsulating package namespace before attempting to load a name from
its environment.
}
\section{Functions}{
\itemize{
\item \code{suggests_double_colon()}: Double-colon operator

\item \code{suggests_triple_colon()}: Triple-colon operator

}}
