% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{feature}
\alias{feature}
\title{Declare a region of code as relating to a feature}
\usage{
feature(features, fallback, env = parent.frame())
}
\description{
Automatically check for the ability to load suggested packages relating
to a specific feature, and choose what to do should the packages be
unavailable.
}
