% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_infixes.R
\name{\%?\%}
\alias{\%?\%}
\title{If not error else}
\usage{
lhs \%?\% rhs
}
\description{
Return the result of the left-hand-side if it does not produce an error,
otherwise return the result of the right-hand-side.
}
