% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import a package as a Suggests dependency}
\usage{
import(pkg, ..., create = FALSE)
}
\arguments{
\item{pkgname}{A \code{character} package name to import as a "Suggests"
dependency}

\item{unavailable_callback}{A \code{function} to be called when the suggested
package is meant to be used, but unavailable.}

\item{env}{A \code{environment} in which the suggested package object should
be assigned. By default, an object of the same name as the package is
created in the parent environment.}
}
\description{
Creates an object in the current environment with the same name as the
package, which can be used just like a package to access its namespace,
throwing an error if the package is not installed.
}
