% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_infixes.R
\name{infix-null-or}
\alias{infix-null-or}
\alias{\%||\%}
\title{If not null else}
\usage{
lhs \%||\% rhs
}
\description{
Return the left-hand-side if it is not \code{NULL}, otherwise return the result of
the right-hand-side.
}
