% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_infixes.R
\name{infix-raise}
\alias{infix-raise}
\alias{\%!\%}
\title{Capture and coerce conditions}
\usage{
lhs \%!\% rhs
}
\description{
Should the left-hand-side emit a condition (for example, an error, warning
or message), capture it and pass it to a right-hand-side function
}
\examples{
(1 + "a") \%!\% warning

}
