% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_infixes.R
\name{infix-try-load-suggests-or}
\alias{infix-try-load-suggests-or}
\alias{\%?s\%}
\title{If not suggests error else}
\usage{
lhs \%?s\% rhs
}
\description{
Return the result of the left-hand-side if it does not produce a suggests
error, otherwise return the result of the right-hand-side.
}
