% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colon_ops.R
\name{materialize_suggests_stub}
\alias{materialize_suggests_stub}
\title{Materialize a package by name}
\usage{
materialize_suggests_stub(pkg, args, env = parent.frame())
}
\arguments{
\item{args}{(\code{list}) Arguments to pass to \code{\link[=new_suggests]{new_suggests()}}.}

\item{env}{(\code{environment}) Where to call \code{\link[=new_suggests]{new_suggests()}} (used for
discovering the encpsulating package namespace).}

\item{pkgs}{(\code{character(1L)}) A package name.}
}
\value{
A materialized \code{suggests_package} object
}
\description{
A suggests package, upon creation, is a minimalist representation containing
only a package name and object location (environment). When it's needed,
we must first \emph{materialize} the object, meaning that we look up the
dependency version requirements and load up any registered fallbacks.
}
\details{
Note that we haven't actually loaded the package yet!
}
\seealso{
Other suggests-lazy-loading: 
\code{\link{new_suggests_stub}()},
\code{\link{try_get_suggests}()}
}
\concept{suggests-lazy-loading}
\keyword{internal}
