% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggested_callbacks.R
\name{suggested_callbacks}
\alias{suggested_callbacks}
\alias{suggests_callback_error}
\alias{suggests_callback_prompt}
\title{Callbacks To Handle Missing Suggests Packages}
\usage{
suggests_callback_error(sug, ...)

suggests_callback_prompt(sug, ...)
}
\description{
These callbacks can be provided to \code{suggested} when declaring a
suggested dependency. If a Suggested package namespace is accessed, the
callback is used when handling the missing dependency. Generally, these are
mitigating or communicative steps to handle the missing package in its
entirety. By contrast, a package author may also define fallback behaviors
for specific namespace objects when the package is otherwise missing.
}
\section{Callbacks}{

\describe{
\item{\code{suggested_callback_error}}{
Throw an error if a suggested package is not installed.
}
}


\describe{
\item{\code{suggested_callback_prompt}}{
Prompt to install a suggested package and resume if installed
}
}
}

