% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggests-package.R
\docType{package}
\name{suggests}
\alias{suggests}
\alias{suggests-package}
\title{suggests: Declare when Suggested Packages are Needed}
\description{
By adding dependencies to the "Suggests" field of a package's
DESCRIPTION file, and then declaring that they are needed within any
dependent functionality, it is often possible to significantly reduce the
number of "hard" dependencies required by a package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/owenjonesuob/suggests}
  \item Report bugs at \url{https://github.com/owenjonesuob/suggests/issues}
}

}
\author{
\strong{Maintainer}: Owen Jones \email{owenjonesuob@gmail.com}

}
\keyword{internal}
