% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggests.R
\name{suggests_loaded}
\alias{suggests_loaded}
\title{Suggests load states}
\usage{
suggests_loaded(sug)
}
\description{
The \code{.__NAMESPACE__.} value, used to represent a loaded package namespace,
can contain a few different sentinal values to indicate different loading
states.
}
\details{
\describe{
\item{\code{environment} (success)}{
The loaded package namespace.
}
\item{\code{NA_integer_} (fail delayed)}{
Load was attempted, but functionality had a fallback method, delay
output until a behavior without a fallback is needed.
}
\item{\code{NULL} (fail)}{
Load was attempted, but failed.
}
}
}
