% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colon_ops.R
\name{try_get_suggests}
\alias{try_get_suggests}
\title{Search for a suggests package namespace object in an environment}
\usage{
try_get_suggests(x, env = parent.frame())
}
\arguments{
\item{x}{A plausible reference to a suggested package namespace}

\item{env}{(\code{environment}) A location in which to search for package
namespace information.}
}
\value{
A \code{suggests_package} object as created with \code{\link[=new_suggests]{new_suggests()}} or
\code{NULL} if no suggests package of that name is declared by the
encapsulating package.
}
\description{
Search for a suggests package namespace object in an environment
}
\seealso{
Other suggests-lazy-loading: 
\code{\link{materialize_suggests_stub}()},
\code{\link{new_suggests_stub}()}
}
\concept{suggests-lazy-loading}
\keyword{internal}
