% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/need.R
\name{need}
\alias{need}
\title{Declare that a package is needed}
\usage{
need(..., ask = interactive())
}
\arguments{
\item{...}{Names of required packages, as character strings.}

\item{ask}{Whether to give the user the option of installing the required
packages immediately.}
}
\value{
Invisibly, the names of any packages which were installed.
}
\description{
Declare that one or more packages are required by subsequent functionality,
and if they're missing, either prompt the user to install them, or exit with
an informative error message.
}
\examples{
\dontrun{

  read_data <- function(path, clean_names = FALSE) {

    # Call need() as early as possible, to avoid wasted work
    if (isTRUE(clean_names))
      suggests::need("janitor")

    output <- utils::read.csv(path)

    if (isTRUE(clean_names))
      output <- janitor::clean_names(output)

    output
  }
 }

}
