% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggests-package.R
\docType{package}
\name{suggests}
\alias{suggests}
\alias{suggests-package}
\title{suggests: Declare when Suggested Packages are Needed}
\description{
By adding dependencies to the "Suggests" field of a package's
DESCRIPTION file, and then declaring that they are needed within any
dependent functionality, you can often significantly reduce the number of
"hard" dependencies (i.e. Depends/Imports) required by your package.

{suggests} aims to make that workflow as minimal and painless as possible,
primarily via the \code{need()} function, which provides a lightweight, simple,
speedy way to prompt your users to install missing suggested packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/owenjonesuob/suggests}
  \item Report bugs at \url{https://github.com/owenjonesuob/suggests/issues}
}

}
\author{
\strong{Maintainer}: Owen Jones \email{owenjonesuob@gmail.com} (\href{https://orcid.org/0000-0002-8410-9585}{ORCID})

}
\keyword{internal}
